ARMTargetMethodUtil.fixupCall32Site ... +16 added, as the first call to VmThreadLocal.tlaSize() ended up in the call to the BASELINE2OPT adapter call and was out by 4 instructions. VMThreadLocal.tlaSize() is returning ZERO this is wrong, so now debugging this.

+++ com.oracle.max.vm.tests/src/test/arm/jtt/ARMV7JTTTest.java
@@ -106,7 +106,7 @@ public class ARMV7JTTTest extends MaxTestCase {
 
         public Args(long lfirst, float fsecond) {
             this.lfirst = lfirst;
-            this.fsecond = fsecond;
+            this.fsecond = (long)fsecond;
         }
